requirements.txt

python-telegram-bot==13.15
Flask==2.0.3
